move backbone related functions into its own backbone_utils file.
